refactor(fio2): move to src directory
