+++ benchmarks/NestedMap.ts
@@ -3,13 +3,13 @@
  */
 import {Suite} from 'benchmark'
 import * as Fluture from 'fluture'
+import {FIO2, interpretSyncFIO2} from '../src/main/FIO2'
+
+import {noop} from '../src/internals/Noop'
 
 import {FIO} from '../src/main/FIO'
 import {defaultRuntime} from '../src/runtimes/DefaultRuntime'
-import {FIO2, interpretSyncFIO2} from './internals/FIO2'
 import {inc} from './internals/Inc'
-
-import {noop} from '../src/internals/Noop'
 import {PrintLn} from './internals/PrintLn'
 
 const suite = new Suite('NestedMap')
@@ -42,7 +42,7 @@ suite
   .add(
     'FIO2',
     (cb: Defer) => {
-      interpretSyncFIO2(fio2, [], noop, () => cb.resolve())
+      interpretSyncFIO2(fio2, undefined, [], noop, () => cb.resolve())
     },
     {defer: true}
   )
